Extract notes header
